fix bug replacing sequences.
When replacing just the sequence items or default state, the current
sequence would be replaced by the new sequence. Because the new sequence
did not have the controller or executor for the peripheral, the
controller and executor would be orphaned. On the next update, a new
controller and executor would be created but the old ones would not be
found for deletion, so there would be multiple controllers and executors
for the same peripheral running simultaneously.